PlantByCountry Full Description

+++ GpscWebApi/Models/LocationModel.cs
@@ -7,7 +7,7 @@ namespace GpscWebApi.Models
 {
     public class LocationModel
     {
-        public double Lat { get; set; }
-        public double Lng { get; set; }
+        public decimal Lat { get; set; }
+        public decimal Lng { get; set; }
     }
 }
